Deployment/requirements.txt
absl-py
aiohttp 
argon2-cffi 
astunparse==1.6.3
async-generator 
async-timeout==3.0.1
attrs 
backcall
bleach
blinker==1.4
brotlipy==0.7.0
cachetools 
certifi==2020.12.5
cryptography==3.2
cffi
chardet 
click 
colorama 
coverage
curl==0.0.1
cycler==0.10.0
Cython==0.29.14
decorator
defusedxml 
entrypoints==0.3
Flask==1.1.2
gast==0.3.3
gensim==3.8.3
grpcio===1.32.0
gunicorn==20.0.4
h5py==2.10.0
imageio==2.9.0
importlib-metadata 
ipykernel 
idna
ipython
ipython-genutils 
itsdangerous==1.1.0
jedi==0.17.0
Jinja2 
joblib==1.0.1
jsonschema 
jupyter-client 
jupyter-core 
jupyterlab-pygments 
Keras
Keras-Applications 
Keras-Preprocessing 
kiwisolver 
Markdown
MarkupSafe==1.1.1
mistune==0.8.4
mkl-fft==1.0.6
multidict 
nbclient 
nbconvert
nbformat 
nest-asyncio 
networkx==2.5
nltk==3.5
notebook 
oauthlib==3.1.0
olefile==0.46
packaging
pandas==1.1.5
pandocfilters
parso
pickleshare 
Pillow
prometheus-client 
prompt-toolkit 
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser
pydot==1.4.2
Pygments 
PyJWT 
pyOpenSSL
pyparsing 
pyreadline==2.1
pyrsistent
PySocks 
python-dateutil 
pytz==2021.1
PyWavelets==1.1.1
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
regex==2020.11.13
requests
requests-oauthlib==1.3.0
retrying==1.3.3
rsa
scikit-learn==0.24.1
Send2Trash 
sip==6.0.3
six
sklearn==0.0
smart-open==4.2.0
tensorboard
tensorboard-plugin-wit
tensorflow-cpu
tensorflow-estimator 
termcolor
terminado
testpath
threadpoolctl
tifffile
tornado 
tqdm
traitlets 
typing-extensions
urllib3 
wcwidth 
webencodings
Werkzeug 
win-inet-pton
wincertstore
wrapt
wheel
yarl
zipp 
